update: Spring JPA Page API 스펙에 맞게 수정

+++ src/main/java/team/incude/gsmc/v2/domain/member/presentation/data/response/SearchStudentResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record SearchStudentResponse(
         Integer totalPage,
-        Integer totalElements,
+        Long totalElements,
         List<GetStudentResponse> results
 ) {
 }
